chore/codygatewayevents: extract into standalone package for reuse
